requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
apache-airflow==2.7.2
apache-airflow-providers-common-sql==1.8.0
apache-airflow-providers-ftp==3.6.0
apache-airflow-providers-http==4.6.0
apache-airflow-providers-imap==3.4.0
apache-airflow-providers-sqlite==3.5.0
apispec==6.3.0
argcomplete==3.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Automat==22.10.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.7.0
blis==0.7.11
cachelib==0.9.0
cachetools==5.3.2
catalogue==2.0.10
cattrs==23.1.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
clickclick==20.10.2
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorama==0.4.6
colorlog==4.8.0
comm==0.1.4
confection==0.1.3
ConfigUpdater==3.1.1
connexion==2.14.2
constantly==23.10.4
contourpy==1.1.1
cron-descriptor==1.4.0
croniter==2.0.1
cryptography==41.0.5
cssselect==1.2.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dask==2023.10.1
databricks-sdk==0.29.0
datasets==2.18.0
DateTime==5.5
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
Django==5.0.6
dnspython==2.6.1
docker==7.1.0
docutils==0.20.1
email-validator==1.3.1
email-verify==0.1.10
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.13.1
Flask==2.2.3
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Moment==1.0.6
Flask-MySQLdb==2.0.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.1
flatbuffers==23.5.26
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
gast==0.5.4
gensim==4.3.2
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.23.4
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
google-re2==1.1
googleapis-common-protos==1.61.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
graphviz==0.20.1
greenlet==3.0.1
grpcio==1.59.2
gunicorn==20.1.0
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.22.2
hyperlink==21.0.0
idna==3.4
imageio==2.34.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itemadapter==0.9.0
itemloaders==1.3.1
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsoncodable==0.1.7
jsonify==0.5
jsonpickle==3.2.2
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.24.0
k-selenium-cookies==0.0.4
kafka-python==2.0.2
kaggle==1.6.12
kagglehub==0.3.3
kazoo==2.5.0
keras==2.14.0
kiwisolver==1.4.5
kproxy==0.0.1
langcodes==3.3.0
lazy-object-proxy==1.9.0
lazy_loader==0.4
libclang==16.0.6
lightgbm==4.1.0
limits==3.6.0
linkify-it-py==2.0.2
locket==1.0.0
lockfile==0.12.2
lxml==4.9.3
Mako==1.2.4
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
MechanicalSoup==1.3.0
mistune==3.0.1
ml-dtypes==0.2.0
mlflow==2.15.0
mlflow-skinny==2.15.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.16
murmurhash==1.0.10
mysql==0.0.3
mysqlclient==2.2.4
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.2.1
nltk==3.8.1
noraise==0.0.16
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.0
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opt-einsum==3.3.0
ordered-set==4.1.0
outcome==1.3.0.post0
overrides==7.4.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.0
parsel==1.9.1
parso==0.8.3
partd==1.4.1
pathspec==0.11.2
patsy==0.5.3
pdf2image==1.17.0
pendulum==2.1.2
pickleshare==0.7.5
Pillow==10.1.0
pipdeptree==2.23.4
platformdirs==3.10.0
plotly==5.21.0
pluggy==1.5.0
preshed==3.0.9
prison==0.2.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
Protego==0.3.1
proto-plus==1.24.0
protobuf==4.24.4
psutil==5.9.5
psycopg2==2.9.5
pure-eval==0.2.2
py3dns==4.0.2
py4j==0.10.9.7
pyarrow==14.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycurl==7.45.3
pydantic==2.10.3
pydantic_core==2.27.1
PyDispatcher==2.0.7
pydot==3.0.3
Pygments==2.16.1
pyhunter==1.7
PyJWT==2.8.0
pykafka==2.8.0
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
PyMySQL==1.1.1
pyOpenSSL==24.1.0
pyparsing==3.1.1
PyPDF2==3.0.1
pyquery==2.0.0
PySocks==1.7.1
pyspark==3.5.0
pyspider==0.3.10
pytest==8.3.3
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.0
python-json-logger==2.0.7
python-nvd3==0.15.0
python-pptx==0.6.23
python-slugify==8.0.1
pytube==15.0.0
pytz==2024.1
pytzdata==2020.1
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.1
querystring-parser==1.2.4
queuelib==1.7.0
redis==5.0.1
referencing==0.30.2
regex==2023.10.3
reportlab==4.1.0
requests==2.28.2
requests-file==2.1.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rich-argparse==1.4.0
rpds-py==0.10.0
rsa==4.9
scikit-image==0.24.0
scikit-learn==1.3.2
scipy==1.11.3
Scrapy==2.11.2
seaborn==0.13.0
selenium==4.15.2
selenium-browser==0.0.15
selenium-chrome==0.0.29
Send2Trash==1.8.2
service-identity==24.1.0
setproctitle==1.3.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.50
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.2
statsmodels==0.14.0
sympy==1.12
tabulate==0.9.0
tblib==3.0.0
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
thinc==8.2.1
threadpoolctl==3.2.0
tifffile==2024.9.20
tinycss2==1.2.1
tk==0.1.0
tldextract==5.1.2
toolz==0.12.0
torch==2.1.0
tornado==6.4.1
tqdm==4.66.1
traitlets==5.9.0
trio==0.23.1
trio-websocket==0.11.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typer==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
u-msgpack-python==2.8.0
uc-micro-py==1.0.2
ucimlrepo==0.0.6
undetected-chromedriver==3.5.5
unicodecsv==0.14.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.19
w3lib==2.2.1
waitress==3.0.0
wasabi==1.1.2
wcwidth==0.2.6
weasel==0.3.3
webcolors==1.13
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.6.2
websockets==12.0
Werkzeug==2.2.3
widgetsnbextension==4.0.9
wrapt==1.14.1
WsgiDAV==4.3.3
wsproto==1.2.0
WTForms==3.1.2
xgboost==2.0.1
XlsxWriter==3.2.0
xpath-utils==0.0.3
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
zope.interface==6.4.post2
